requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
openai
python-dotenv
PyMuPDF
psycopg2-binary
